Extract param types into their own namespace
