+++ lib/compendium/param_types/dropdown.rb
@@ -1,0 +1,9 @@
+module Compendium
+  module ParamTypes
+    class Dropdown < WithChoices
+      def dropdown?
+        true
+      end
+    end
+  end
+end
