+++ lib/compendium/param_types/radio.rb
@@ -1,0 +1,9 @@
+module Compendium
+  module ParamTypes
+    class Radio < WithChoices
+      def radio?
+        true
+      end
+    end
+  end
+end
