+++ lib/compendium/param_types/scalar.rb
@@ -1,0 +1,14 @@
+module Compendium
+  module ParamTypes
+    class Scalar < Param
+      def initialize(obj, *)
+        super obj
+      end
+
+      # A scalar param just keeps track of a value with no modifications
+      def scalar?
+        true
+      end
+    end
+  end
+end
