+++ spec/compendium/collection_query_spec.rb
