+++ spec/compendium/context_wrapper_spec.rb
