+++ spec/compendium/count_query_spec.rb
