+++ spec/compendium/dsl_spec.rb
