+++ spec/compendium/metric_spec.rb
