+++ spec/compendium/option_spec.rb
