+++ spec/compendium/params_spec.rb
