+++ spec/compendium/query_spec.rb
