+++ spec/compendium/report_spec.rb
