+++ spec/compendium/result_set_spec.rb
