+++ spec/compendium/sum_query_spec.rb
