+++ spec/compendium/through_query_spec.rb
